Define a TypedPropertyMatcher and a LegacyPropertyMatcher
The TypedPropertyMatcher will rely on type checking to perform the matching. The legacy matcher will perform the current name-based comparison on the type object.

All current usages of the PropertyMatcher are refactored to use LegacyPropertyMatcher to preserve the current behavior.

Rules can progressively be opted in the typed property matcher by setting the `useTypedPropertyMatcher` in their `.PatternEngineConfig`.

PiperOrigin-RevId: 783766684
Change-Id: I62d517c7c57522c0f23ff1525677cceb7cd8e7aa