Add an option to ignore types that matches because they are structurally compatible with a banned type.
Type checking is performed structurally. This means that if A.foo is banned, but B is assignable to A, B.foo will be banned. This is generally WAI with B extends A and foo has the same semantics (e.g. HTMLElement.innerHTML is banned, and this bans HTMLDivElement.innerHTML). However this can raise false positives when interfaces are close but not semantially related. For instance, HTMLLinkElement is assignable to HTMLStyleElement, but HTMLLinkElement.textContent is not related to HTMLStyleElement#textContent (which we want to ban). This change lets allowlist types that are know to cause false positives because they are structurally compatible with the banned type.

PiperOrigin-RevId: 811739569
Change-Id: I8a63b84eb55556ec4d004e20c4b5cbab1203acd3